Added a air alarm to the Processing Room.
Added a pistol to the Captain's safe, as a last resort.

Added the Kellion Technician outfit to the code.

Added a standard device power cell to the Supply Officer's locker to power the Stick of Discipline.

Assigned Line Cook and Bartender only accesses and added them to the appropriate doors.

Changed the Captain's safe to a modern, wall-safe.

The Monitoring Room door bolt button now works.

Fixed the door name on the janitorial closet, south of Cargo.

Fixed the windows on the Aegis Colony Security Checkpoint to be the correct type of windows.

Fixed the Kellion Engineer spawning with Unitologist gear.